Rename test project and update namespaces for consistency
