+++ test/LayeredCraft.StructuredLogging.Tests/LayeredCraft.StructuredLogging.Tests.csproj
@@ -4,11 +4,18 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <OutputType>Exe</OutputType>
-    <RootNamespace>LayeredCraft.StructuredLogging.Test</RootNamespace>
     <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <LangVersion>default</LangVersion>
-    <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
-    <UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
+    <!--
+    To enable the Microsoft Testing Platform 'dotnet test' experience, add property:
+      <TestingPlatformDotnetTestSupport>true</TestingPlatformDotnetTestSupport>
+
+    To enable the Microsoft Testing Platform native command line experience, add property:
+      <UseMicrosoftTestingPlatformRunner>true</UseMicrosoftTestingPlatformRunner>
+
+    For more information on Microsoft Testing Platform support in xUnit.net, please visit:
+      https://xunit.net/docs/getting-started/v3/microsoft-testing-platform
+    -->
   </PropertyGroup>
 
   <ItemGroup>
@@ -31,10 +38,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\LayeredCraft.StructuredLogging\LayeredCraft.StructuredLogging.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include=".github\" />
   </ItemGroup>
 
 </Project>
