+++ test/LayeredCraft.StructuredLogging.Tests/xunit.runner.json
